+++ Aponus Web API/Models/EstadosSuministrosMovimientosStock.cs
@@ -8,7 +8,7 @@ namespace Aponus_Web_API.Models
     {
         [Key]
         [Column(("ID_ESTADO"))]
-        public byte IdEstado { get; set; } = 1;
+        public int IdEstado { get; set; } = 1;
 
         [Column("DESCRIPCION")]
         public string Descripcion { get; set; }
