+++ Aponus Web API/Models/ProductosDescripcion.cs
@@ -13,7 +13,7 @@ public partial class ProductosDescripcion
     public string? DescripcionProducto { get; set; }
 
     [ForeignKey("ID_ESTADO")]
-    public byte IdEstado { get; set; }
+    public int IdEstado { get; set; }
     public virtual EstadosProductosDescripciones IdEstadoNavigation { get; set; } = new EstadosProductosDescripciones();
     public virtual ICollection<Producto> Productos { get; set; } = new List <Producto>();
    // public virtual ICollection<Productos_Tipos_Descripcion> Producto_Tipo_Descripcione { get; set; } = new List<Productos_Tipos_Descripcion>();
